use git timeout setting on agent.jar directly
this appears to be the only way to get the gerrit plugin
to increase the timeout for when it's fetching things.

still don't know if this will solve the issues as it's
not clear if a hung gerrit ssh will actually recover or if
it's permanently blocked, which means this fix would make
things worse if so.