set timeout down to 5.
30 didn't help as the issue is apparently a total hang.

instead we have re-read the gerrit plugin instructions and
have selected a different git strategy that correctly goes
to just the rev we need.  this is expected to solve the issue
as we no longer are fetching all refs.   this is essentially
what the pep484 and pep8 jobs do (and really we could just do
git commandline in the build steps for everything if we wanted).