things are running
lots of glitches now due to jenkins upgrades w/ ssh
stuff and whatnot.   Can't get Oracle to run on zotac; if I run the
steps manually in a container, everything works.  try to turn
the container into an image, not working.  building in buildah,
the build just fails without a clear reason.

however oracle is running on d3 and seems likely will run on dell,
so maybe zotac just doesn't have enough RAM, so will keep it off
the "heavy" label for now.